+++ protarrow/common.py
@@ -9,10 +9,18 @@ M = TypeVar("M", bound=Message)
 SUPPORTED_ENUM_TYPES = (
     pa.int32(),
     pa.binary(),
+    pa.large_binary(),
     pa.string(),
+    pa.large_string(),
     pa.dictionary(pa.int32(), pa.string()),
+    pa.dictionary(pa.int32(), pa.large_string()),
     pa.dictionary(pa.int32(), pa.binary()),
+    pa.dictionary(pa.int32(), pa.large_binary()),
 )
+
+SUPPORTED_LIST_ARRAY_TYPES = (pa.ListArray, pa.LargeListArray)
+SUPPORTED_STRING_TYPES = (pa.string(), pa.large_string())
+SUPPORTED_BINARY_TYPES = (pa.binary(), pa.large_binary())
 
 
 @dataclasses.dataclass(frozen=True)
@@ -28,16 +36,29 @@ class ProtarrowConfig:
     list_value_name: str = "item"
     map_value_name: str = "value"
     field_number_key: Optional[bytes] = None
+    string_type: pa.DataType = pa.string()
+    binary_type: pa.DataType = pa.binary()
+    list_array_type: type = pa.ListArray
 
     def __post_init__(self):
         assert self.enum_type in SUPPORTED_ENUM_TYPES
         assert isinstance(self.field_number_key, (bytes, type(None)))
+        assert self.string_type in SUPPORTED_STRING_TYPES
+        assert self.binary_type in SUPPORTED_BINARY_TYPES
+        assert self.list_array_type in SUPPORTED_LIST_ARRAY_TYPES
 
     def field_metadata(self, field_number: int) -> Optional[dict[bytes, bytes]]:
         if self.field_number_key is None:
             return None
         else:
             return {self.field_number_key: str(field_number).encode("utf-8")}
+
+    def list_(self, item_type: pa.DataType) -> pa.DataType:
+        return (pa.list_ if self.list_array_type is pa.ListArray else pa.large_list)(
+            pa.field(
+                self.list_value_name, item_type, nullable=self.list_value_nullable
+            ),
+        )
 
 
 def is_binary_enum(data_type: pa.DataType) -> bool: